Split methods logically, add yard comments and hide unnecessary public methods
