+++ lib/ams_lazy_relationships/loaders/simple_belongs_to.rb
@@ -18,8 +18,6 @@ module AmsLazyRelationships
         @foreign_key = foreign_key.to_sym
       end
 
-      attr_reader :association_class_name, :foreign_key
-
       # Lazy loads and yields the data when evaluating
       # @param record [Object] an object for which we're loading the belongs to data
       # @param block [Proc] a block to execute when data is evaluated
@@ -35,6 +33,8 @@ module AmsLazyRelationships
       end
 
       private
+
+      attr_reader :association_class_name, :foreign_key
 
       def load_data(records)
         data_ids = records.map(&foreign_key).compact.uniq
